Calculate CCRI using negative exponential for given parameters This function calculates CCRI using negative exponential for given parameters based on provided metrics and parameters.

Calculate CCRI using negative exponential for given parameters This function calculates CCRI using negative exponential for given parameters based on provided metrics and parameters.

ccri_neg_exponential_function(
  dispersal_parameter_gamma_val,
  link_threshold,
  distance_matrix = the$distance_matrix,
  thresholded_crop_values,
  crop_raster,
  crop_cells_above_threshold,
  betweenness_metric = FALSE,
  node_strength = FALSE,
  sum_of_nearest_neighbors = FALSE,
  eigenvector_centrality = FALSE
)

Arguments

dispersal_parameter_gamma_val

A list of gamma values

link_threshold

A threshold value for link

distance_matrix

distance matrix calculated during initialize_crop_data().

thresholded_crop_values

crop values above threshold

crop_raster

A raster object for crop raster

crop_cells_above_threshold

A list of crop cells above threshold

betweenness_metric

A boolean value to calculate betweenness metric

node_strength

A boolean value to calculate node strength

sum_of_nearest_neighbors

A boolean value to calculate sum of nearest neighbors

eigenvector_centrality

A boolean value to calculate eigenvector centrality

Value

A list of calculated CCRI values using negative exponential