Calculate mean of raster objects

Wrapper for terra::mean(). Calculates mean of list of rasters.

Usage

hci_mean(
  indices,
  global = FALSE,
  east = NULL,
  west = NULL,
  geoscale = NULL,
  res = reso(),
  plt = TRUE,
  outdir = tempdir()
)

Arguments

indices

List of SpatRasters. This input represents the spatial raster collection for which mean is to be calculated.

global

Logical. TRUE if global analysis is desired, FALSE otherwise. east and west are required when TRUE.

east

SpatRaster. Collection of risk indices on eastern extent.

west

SpatRaster. Collection of risk indices on western extent. When TRUE, geoscale is ignored. Default is TRUE.

geoscale

Vector. geographical scale. Default is NULL.

res

Numeric. Map resolution. This value is used in aggregation and dis-aggregation operation. Default is reso().

plt

TRUE if need to plot mean map, FALSE otherwise.

outdir

Character. Output directory for saving raster in TIFF format. Default is tempdir().

Value

RiskMap. Contains result in the form of SpatRaster objects and file path of the saved maps.